[kong]: move url parameters extractor to Kontrol trait
